requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
psycopg==3.1.18
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rpds-py==0.18.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
